map_tools/requirements.txt
pygame_ce==2.3.2
pygame_gui==0.3.9